app/requirements.txt
langchain==0.2.16
langchain-community==0.2.17
langchain-core==0.2.41
langchain-openai==0.1.25
langchain-text-splitters==0.2.4
huggingface-hub==0.25.1
numpy==1.26.4
openai==1.51.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
datasets==3.0.1
pandas==2.2.3
python-dotenv==1.0.1
streamlit==1.34.0
tqdm==4.66.5
pydantic==2.9.2
pydantic-core==2.23.4
pymongo==4.10.1
semantic-router==0.0.71
streamlit-browser-session-storage==0.0.10
guardrails-ai==0.5.15
guardrails-api==0.0.3
guardrails-api-client==0.3.13
guardrails-hub-types==0.0.4
